requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
numpy==1.24.3
scikit-learn==1.2.2
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.6
